# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SERVED. On-Demand Process Serving Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /demo/invites/{token}:
    get:
      operationId: getDemoInvite
      tags: [demo]
      summary: Look up an attorney-demo invite by token (public)
      description: |
        Public, unauthenticated. Returns minimal metadata about the invite
        so the watch page can render a tailored email-gate. Never returns
        the recipient email itself — the visitor must type it back in.
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Invite found and still valid.
          content:
            application/json:
              schema:
                type: object
                required: [validUntil, alreadyConfirmed]
                properties:
                  recipientName:
                    type: string
                    nullable: true
                  validUntil:
                    type: string
                    format: date-time
                  alreadyConfirmed:
                    type: boolean
        "404":
          description: Token not found, expired, or revoked.
  /demo/invites/{token}/confirm:
    post:
      operationId: confirmDemoInvite
      tags: [demo]
      summary: Soft email-gate for the attorney-demo watch page (public)
      description: |
        Public, unauthenticated. The visitor types the email the invite
        was addressed to; on a case-insensitive match we stamp viewed_at
        (idempotent) and return ok. Wrong email returns 403 without
        leaking the real address.
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        "200":
          description: Email matches; viewed_at recorded.
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok:
                    type: boolean
        "400":
          description: Email missing or malformed.
        "403":
          description: Email does not match the invitation.
        "404":
          description: Token not found, expired, or revoked.
...